+++ docs/change_log.rst
@@ -1,7 +1,9 @@
 更新记录
 ===========================
-2018.12.28 v1.1.5
+2018.12.29 v1.1.5
 ---------------------------
+* `#173 <https://github.com/chinapnr/fishbase/issues/173>`_, random, add function :meth:`fish_random.get_random_zone_name`, doc and unittest;
+* `#162 <https://github.com/chinapnr/fishbase/issues/162>`_, random, add function :meth:`fish_random.gen_float_by_range`, doc and unittest;
 * `#166 <https://github.com/chinapnr/fishbase/issues/166>`_, random, add function :meth:`fish_random.gen_mobile`, doc and unittest;
 * `#171 <https://github.com/chinapnr/fishbase/issues/171>`_, random, add function :meth:`fish_random.gen_name`, doc and unittest;
 * `#163 <https://github.com/chinapnr/fishbase/issues/163>`_, random, add function :meth:`fish_random.gen_string_by_range`, doc and unittest;